+++ rusty_fractals/src/constants.rs
@@ -8,21 +8,21 @@ pub const NEIGHBOURS: u8 = 3;
 pub const PATH: &str = "/Fractals/";
 
 // to remove noise, ignore pixels with this value or less as 0
-pub const COLORING_THRESHOLD: u32 = 3;
+pub const COLORING_THRESHOLD: u64 = 3;
 
 // Delete paths shorter than this
 // Remembered paths got shorter as some of their elements moved out of displayed Area
-pub const MINIMUM_PATH_LENGTH: u32 = 4;
+pub const MINIMUM_PATH_LENGTH: u64 = 4;
 
 // 4 is quadrance from (0, 0)
 // If intermediate calculation result [re,im] spirals beyond this boundary. Calculation stops as divergent.
-pub const CALCULATION_BOUNDARY: u32 = 4;
+pub const CALCULATION_BOUNDARY: u64 = 4;
 
 // 0.98 is a good choice for 25fps and moderate speed
 pub const ZOOM: f64 = 0.98;
 
 // take result data snapshot for comparison at well colored frame
-pub const TAKE_MEASURES_AT_FRAME: u32 = 20;
+pub const TAKE_MEASURES_AT_FRAME: u64 = 20;
 
 // phoenix fractal constants
 pub const PHOENIX_INIT_C: f64 = 0.35;
